fn: konvert is discovered in items if not provided
When running konvert as a KRM function, it will attempt to discover a
Konvert-file as the functionConfig if not otherwise specified. This
enables some workflows with Kpt. When the Konvert-file is specified
explicitly as a functionConfig, it is considered a
meta-resource (https://kpt.dev/book/04-using-functions/02-imperative-function-execution?id=mutate-meta-resources)
and is not considered a target for other KRM functions (without using
`--include-meta-resources`) like `apply-setters`. And
`--include-meta-resources` is not really an option yet (I don't think)
in declarative pipelines.